Flight Track Planning tool

makeTP generates a flight track chaining up point objects with respect to a heterogenous surface and known obstacles as documented by an DSM for taking top down pictures. It creates a single control file for autonomous picture retrieval flights.

makeTP(
  projectDir = tempdir(),
  locationName = "treePos",
  missionTrackList = NULL,
  launchPos = c(8.772055, 50.814689),
  demFn = NULL,
  flightAltitude = 100,
  climbDist = 7.5,
  aboveTreeAlt = 15,
  circleRadius = 1,
  takeOffAlt = 50,
  presetFlightTask = "remote",
  maxSpeed = 25,
  followSurfaceRes = 5,
  altFilter = 0.5,
  windCondition = 1,
  launchAltitude = -9999,
  uavType = "pixhawk",
  cameraType = "MAPIR2",
  copy = FALSE,
  runDir = ""
)

Arguments

projectDir

character path to the main folder where several projects can be hosted, default is tempdir()

locationName

character base name string of the mission, default is "treePos"

missionTrackList

character filename of the mission tracklist (target positions), default is NULL

launchPos

list launch position c(longitude,latitude), default is c(8.772055,50.814689)

demFn

character filename of the used DSM data file, default is NULL

flightAltitude

numeric set the AGL flight altitude (AGL while the provided raster model represents this surface) of the mission, default is 100 default is (= 0.0). If set to -99 it will be calculated from the swath width of the pictures. NOTE: This makes only sense for followingTerrain = TRUE to smooth curves. For flightPlanMode = "waypoint" camera actions (DJI only EXPERIMENTAL) are DISABLED during curve flights.

climbDist

numeric distance within the uav will climb on the caluclated save flight altitude in meter, default is 7.5

aboveTreeAlt

numeric minimum flight height above target trees in meter, default is 15.0

circleRadius

numeric radius to circle around above target trees in meter, default is 1.0

takeOffAlt

altitude numeric climb altitude of the uav at take off position in meter, default is 50.0

presetFlightTask

character (DJI only EXPERIMENTAL). NOTE: it is strongly recommended to use the default "remote"
Further options are:
"simple_ortho" takes one picture/waypoint, "multi_ortho" takes 4 picture at a waypoint, two vertically down and two in forward and backward viewing direction and an angele of -60deg, "simple_pano" takes a 360 deg panorama picture and "remote" which assumes that the camera is controlled by the remote control (RC)

maxSpeed

numeric cruising speed, default is 25.0

followSurfaceRes

numeric, default is 5 meter.

altFilter

numeric allowed altitude differences bewteen two waypoints in meter, default is 0.5

windCondition

numericoptions are 1= calm 2= light air 1-5km/h, 3= light breeze 6-11km/h, 4=gentle breeze 12-19km/h 5= moderate breeze 20-28km/h, default is 1

launchAltitude

numeric altitude of launch position. If set to -9999 a DEM is required for extracting the MSL, default is -9999

uavType

character type of UAV. Currently "dji_csv" and "pixhawk" are supported, default is "pixhawk"

cameraType

character, default is "MAPIR2".

copy

boolean copy used file to data folder default is FALSE

runDir

character runtime folder

Examples

if (FALSE) {
## (1) get example DEM data
dsmFn <- system.file("extdata", "mrbiko.tif", package = "uavRmp")
## (2) make position flight plan
makeTP  <-  makeTP(missionTrackList= tutorial_flightArea,
                  demFn = dsmFn,
                  uavType = "pixhawk",
                  launchPos = c(8.679,50.856))
}