Upload MAV compliant mission File to a 3DR Solo

solo_upload provides a crude interface to upload the Solo mission file to the 3dr SOLO

solo_upload(
  missionFile = NULL,
  connection = "udp:10.1.1.166:14550",
  prearm = "-1"
)

Arguments

missionFile

mission file to upload

connection

a valid connection string to the Solo default is "udp:10.1.1.166:14550"

prearm

character controls the prearm status of the Solo prearm check
0=Disabled
1=Enabled
-3=Skip Baro
-5=Skip Compass
-9=Skip GPS
-17=Skip INS
-33=Skip Params/Rangefinder
-65=Skip RC
127=Skip Voltage
default is -1

Find more information at prearm safety,
Mission import export script.

Note

Becareful with fooling around with the prearm stuff. It is kind of VERY sensitive for the later autonomous flights!
For using the Solo stuff you need to install:
sudo pip install pymavlink;
sudo pip install dronekit-sitl;
sudo pip install dronekit;
sudo apt-get install sshpass
Additionally you need to be connected to a running 3DR Solo uav

Examples

wp <- system.file("extdata", "MAVLINK_waypoints.txt", package = "uavRmp")
if (FALSE) {
solo_upload( missionFile = wp)
}